export components into files
